Time: 23 ms (6.42%), Space: 8.7 MB (14.09%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int mySqrt(int x) {
         if (x == 2) return 1;
-        long i = 1;
+        long long i = 1;
         while ((i*i) != x) {
             if ((i*i) == x) {
                 break;
